backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OpenAI client
openai>=1.50.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2

# Optional: RBC Security (only available in RBC environment)
# rbc_security  # Uncomment when deploying to RBC environment
